tests/wpt/tests/focus/activeelement-after-immediately-focusing-different-site-iframe.html

<!doctype html>
<meta charset=utf-8>
<title>activeElement when immediately focusing different-site iframe</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script>
setup({explicit_done:true});
window.onmessage = function(e) {
    test(function() {
        assert_equals(e.data, "outerparser,activeElement:BODY,willfocusiframe,didfocusiframe,activeElement:IFRAME,willbluriframe,didbluriframe,activeElement:BODY,willspineventloop,", 'Check log');
    }, "Check result");
    w.close();
    done();
};
var w = window.open("support/activeelement-after-immediately-focusing-different-site-iframe-outer.sub.html");
</script>
