LayoutTests/imported/w3c/web-platform-tests/focus/activeelement-after-immediately-focusing-different-site-iframe-expected.txt


FAIL Check result assert_equals: Check log expected "outerparser,activeElement:BODY,willfocusiframe,didfocusiframe,activeElement:IFRAME,willbluriframe,didbluriframe,activeElement:BODY,willspineventloop," but got "outerparser,activeElement:BODY,willfocusiframe,didfocusiframe,activeElement:IFRAME,willbluriframe,didbluriframe,activeElement:IFRAME,willspineventloop,"
